backend/requirements.txt
fastapi
uvicorn[standard]
sentry-sdk
python-dotenv
httpx
